moved drizzle stuff into db directory
