+++ package.json
@@ -30,73 +30,73 @@
   },
   "dependencies": {
     "@focus-reactive/sanity-plugin-inline-svg-input": "^1.3.0",
-    "@marsidev/react-turnstile": "^1.4.2",
-    "@neondatabase/serverless": "^1.0.1",
+    "@marsidev/react-turnstile": "^1.5.0",
+    "@neondatabase/serverless": "^1.0.2",
     "@sanity/icons": "^3.7.4",
-    "@sanity/image-url": "^2.0.3",
-    "@upstash/redis": "^1.35.4",
+    "@sanity/image-url": "^2.1.1",
+    "@upstash/redis": "^1.37.0",
     "argon2": "^0.44.0",
-    "bufferutil": "^4.0.9",
+    "bufferutil": "^4.1.0",
     "clsx": "^2.1.1",
-    "dotenv": "^17.2.2",
-    "drizzle-orm": "0.45.1",
+    "dotenv": "^17.3.1",
+    "drizzle-orm": "0.45.2",
     "drizzle-zod": "^0.8.3",
     "embla-carousel-react": "^8.6.0",
-    "next": "^16.1.6",
-    "next-cloudinary": "^6.16.0",
-    "next-sanity": "^12.1.0",
-    "pg": "^8.16.3",
+    "next": "^16.2.2",
+    "next-cloudinary": "^6.17.5",
+    "next-sanity": "^12.2.1",
+    "pg": "^8.20.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
-    "sanity": "^5.12.0",
+    "sanity": "^5.18.0",
     "sanity-plugin-cloudinary": "^1.4.1",
     "server-only": "^0.0.1",
-    "styled-components": "^6.1.19",
-    "tailwind-merge": "^3.3.1",
-    "ws": "^8.18.3",
-    "zod": "^4.1.11"
+    "styled-components": "^6.3.12",
+    "tailwind-merge": "^3.5.0",
+    "ws": "^8.20.0",
+    "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^5.0.1",
-    "@eslint/js": "^9.36.0",
-    "@happy-dom/global-registrator": "^20.8.3",
+    "@chromatic-com/storybook": "^5.1.1",
+    "@eslint/js": "^9.39.4",
+    "@happy-dom/global-registrator": "^20.8.9",
     "@sanity/color-input": "^6.0.4",
     "@sanity/eslint-config-studio": "^6.0.0",
-    "@sanity/vision": "^5.12.0",
-    "@storybook/addon-a11y": "10.2.14",
-    "@storybook/addon-docs": "10.2.14",
-    "@storybook/addon-onboarding": "10.2.14",
-    "@storybook/nextjs-vite": "10.2.14",
-    "@tailwindcss/postcss": "^4.1.12",
-    "@testing-library/jest-dom": "^6.7.0",
+    "@sanity/vision": "^5.18.0",
+    "@storybook/addon-a11y": "10.3.3",
+    "@storybook/addon-docs": "10.3.3",
+    "@storybook/addon-onboarding": "10.3.3",
+    "@storybook/nextjs-vite": "10.3.3",
+    "@tailwindcss/postcss": "^4.2.2",
+    "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "16.3.2",
     "@testing-library/user-event": "^14.6.1",
-    "@types/bun": "^1.3.10",
+    "@types/bun": "^1.3.11",
     "@types/jest-axe": "^3.5.9",
-    "@types/jsdom": "^28.0.0",
-    "@types/pg": "^8.15.5",
+    "@types/jsdom": "^28.0.1",
+    "@types/pg": "^8.20.0",
     "@types/react-dom": "^19.2.3",
     "@types/ws": "^8.18.1",
-    "axe-core": "^4.10.3",
-    "bun-types": "^1.2.22",
-    "drizzle-kit": "^0.31.4",
+    "axe-core": "^4.11.2",
+    "bun-types": "^1.3.11",
+    "drizzle-kit": "^0.31.10",
     "drizzle-seed": "^0.3.1",
     "eslint": "9.36.0",
-    "eslint-config-next": "16.1.6",
+    "eslint-config-next": "16.2.2",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-drizzle": "^0.2.3",
     "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jest": "^29.0.1",
-    "eslint-plugin-storybook": "10.2.14",
+    "eslint-plugin-jest": "^29.15.1",
+    "eslint-plugin-storybook": "10.3.3",
     "husky": "^9.1.7",
     "jest-axe": "^10.0.0",
-    "lint-staged": "^16.2.0",
+    "lint-staged": "^16.4.0",
     "postcss": "^8.5.8",
-    "prettier": "^3.6.2",
+    "prettier": "^3.8.1",
     "prettier-plugin-tailwindcss": "^0.7.2",
-    "storybook": "10.2.14",
-    "tailwindcss": "^4.2.1",
-    "typescript-eslint": "^8.56.1"
+    "storybook": "10.3.3",
+    "tailwindcss": "^4.2.2",
+    "typescript-eslint": "^8.58.0"
   },
   "lint-staged": {
     "*.(json|js|jsx|ts|tsx|html|css|md)": [
