+++ src/db/seeds/location/recipients.ts
@@ -1,6 +1,6 @@
 import { seed } from 'drizzle-seed';
-import { db } from '@/_drizzle/db';
-import { recipientsTable } from '@/_drizzle/schemas/recipientsDrizzle';
+import { db } from '@/db/db';
+import { recipientsTable } from '@/db/schemas/recipientsDrizzle';
 
 export const seedRecipients = async () => {
   await seed(db, { recipientsTable }, { seed: 1 }).refine(funcs => ({
